Move BIHData to separate file
